requirements.txt
streamlit
opencv-python
torch
torchvision
numpy
Pillow
transformers
sentencepiece
TTS
gtts
symspellpy
jiwer
matplotlib
seaborn
fasttext